DIGITAL RESUMÉ - INTRO TO WEB
This page includes the following content:

My name and an appropriate description.

An introductory paragraph about myself.

A list of my skills.

A list of my interests.

A list of work experiences.

A list of my education and qualifications.

A circular image of myself.

My contact information.

A navigation menu.